+++ src/plugins/HinderedRotorCM1D.cpp
@@ -25,7 +25,7 @@ namespace mesmer
     virtual bool ReadParameters(gDensityOfStates* gdos, PersistPtr ppDOSC=NULL);
 
     // Provide a function to define particular counts of the DOS of a molecule.
-    virtual bool countCellDOS(gDensityOfStates* mol, int MaximumCell);
+    virtual bool countCellDOS(gDensityOfStates* mol, size_t MaximumCell);
 
     // Provide a function to calculate contribution to canonical partition function.
     // (Mostly for testing purposes.)
@@ -223,7 +223,7 @@ namespace mesmer
   // Calculate classical mechanical 1D rotor densities of states of a free 
   // rotor and convolve them with the main density of states.
   //
-  bool HinderedRotorCM1D::countCellDOS(gDensityOfStates* pDOS, int MaximumCell)
+  bool HinderedRotorCM1D::countCellDOS(gDensityOfStates* pDOS, size_t MaximumCell)
   {
     vector<double> cellDOS;
     if(!pDOS->getCellDensityOfStates(cellDOS, 0, false)) // retrieve the DOS vector without recalculating
